improved test cases: moved common hssf-xssf test to org.apache.poi.ss namespace, removed duplicate tests, refactored code to throw same exceptions, etc
git-svn-id: https://svn.apache.org/repos/asf/poi/trunk@759112 13f79535-47bb-0310-9956-ffa450edef68

+++ src/ooxml/interfaces-jdk15/org/apache/poi/ss/usermodel/Workbook.java
@@ -97,7 +97,7 @@ public interface Workbook {
      * @see Sheet#setSelected(boolean)
      * @param index the index of the sheet to select (0 based)
      */
-    void setSelectedTab(short index);
+    void setSelectedTab(int index);
 
     /**
      * Set the sheet name.
